[refactor/#155] pattern 파일 분리
