+++ src/lib/data/constants.ts
@@ -14,6 +14,10 @@ export const SIGN_UP_ERROR_MESSAGE = {
   title: '회원 가입 실패',
 }
 
+export const SAVE_ERROR_MESSAGE = {
+  title: '저장 실패',
+}
+
 export const TEST_ID_APPLICANT = {
   email: 'testApplicant@email.com',
   password: '00000000',
@@ -22,14 +26,4 @@ export const TEST_ID_APPLICANT = {
 export const TEST_ID_OWNER = {
   email: 'testOwner@email.com',
   password: '00000000',
-}
-
-export const emailPattern = {
-  value: /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/,
-  message: '이메일 형식이 아닙니다.',
-}
-
-export const passwordPattern = {
-  value: /^.{6,}$/,
-  message: '비밀번호는 6자 이상이어야 합니다.',
 }
